+++ RELEASE-NOTES.md
@@ -23,6 +23,7 @@
 * #537 Split JAX-RS interface to Agrest in a separate module
 * #538 "agrest-jaxrs3" module 
 * #541 Remove deprecated metadata responses
+* #549 OpenAPI to use format="partial-time" for time attributes
 * #557 Upgrade Jackson to 2.13.2.x
 * #558 JacksonService swallows parse exception cause
 * #561 Protocol changes - v1.2